Add: Metadata to Event Model
Add: Metadata to Event Model
Update: Unit Test
Update: README
Update: .gitignore
Bump: version to 1.1.0
